Distribute old broadcast tests to correct modules
